split s3 methods into another file, changed static file retrieval for s3
